ITS ALIVE!!!!
Just about. The AI is not functioning. It take too long to make a decision currently although I intend to thread some of the tree work so hopefully that will speed it up. Additionally it is possible that the number of simulations/lookahead depth is too long. Still, it works.