Large updates and improvements to the AI
The AI was previously very very very buggy and whilst it sometimes provided results that looked right under the hood not a lot was going right for it.

This is now better. The AI plays a lot more intelligently (although it still struggles in places such as for some reason often avoiding capturing my peices and dropping peices weirdly).